SCHED-572: Major cleanup of resource package.
