+++ scheduler/services/resource/__init__.py
@@ -1,7 +1,9 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from .base import *
-from .ocs_resource_service import *
+from .event_generators import *
+from .file_resource_service import *
 from .filters import *
-
+from .night_configuration import *
+from .ocs_resource_service import *
+from .resource_service import *
