+++ scheduler/services/resource/data/README.md
@@ -1,0 +1,7 @@
+# Resource Data
+
+This directory contains static OCS resource data files.
+
+Files here are preprocessed using scripts to put them into a standardized format.
+
+These should be removed from the Scheduler repository and made available elsewhere.
